👷(ci) add CI workflow to build Moshi server image
Introduce a basic CI workflow that builds the Moshi server image for L40s GPUs.

In the longer term, if we add support for additional GPU architectures,
we can extend the CI to include dedicated jobs that build the image
with the appropriate CUDA_COMPUTE_CAP value for each GPU type.

Inspired from lasuite/whisper-api CI.